user can now add additional options for compilation

+++ html/compile.php
@@ -1,26 +1,35 @@
 <?php
-
 	error_reporting(E_ALL);
 	ini_set('display_errors', '1');
-
 
 	$c_path = '/tmp/temp.c';
 	$asm_path = '/tmp/temp.s';
 
-//	var_dump($_POST);
-//	echo $_POST['code'];
-
 	/* Compile the file and return the asm code */
 	file_put_contents($c_path, $_POST['code']);
 
-	//echo "gcc $c_path -S -o $asm_path";
-	exec("gcc $c_path -S -masm=intel -o $asm_path");
+	$compiler = $_POST['compiler'];
+	$additional = $_POST['additional'];
 
-	$asm_code = file_get_contents($asm_path);
+	echo $additional . PHP_EOL;
 
-	echo $asm_code;
+	/* Create the command line string, validate it and execute it  */	
+	$cmdline_string = "$compiler $c_path -S -o $asm_path -fno-asynchronous-unwind-tables " . $additional;	
+	$cmdline_string = escapeshellcmd($cmdline_string);	
+	exec($cmdline_string, $exec_output, $ret_code);
 
-	/* Delete the temporary files */
+
+	var_dump($exec_output);
+	
+	if ($ret_code !== 0) {
+		echo "Compilation failed";
+	} else {
+		$asm_code = file_get_contents($asm_path);
+
+		echo $asm_code;
+
+		/* Delete the temporary files */
+		unlink($asm_path);
+	}
 	unlink($c_path);
-	unlink($asm_path);
 ?>
